requirements.txt
python-telegram-bot==13.7
gql==3.0.0a6
openai==0.27.0